改进cur_mov架构
引入modules模式构建cur_mov，以便使用reload动态装载“策略”算法。
